backend/auth_service/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
supabase==2.3.0
pydantic==2.5.3
